# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Vessl API Documentation",
        "title": "Vessl API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/deployments/{id}/logs": {
            "get": {
                "description": "GetLogs endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployments"
                ],
                "summary": "GetLogs endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/deployments/{id}/rollback": {
            "post": {
                "description": "Rollback endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployments"
                ],
                "summary": "Rollback endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/domains/{id}": {
            "delete": {
                "description": "Delete Domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Delete Domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Domain ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/environments/{id}/canvas": {
            "get": {
                "description": "GetEnvironmentCanvas endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Environments"
                ],
                "summary": "GetEnvironmentCanvas endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/projects/{id}/deploy": {
            "post": {
                "description": "DeployProject endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "DeployProject endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/projects/{id}/domains": {
            "get": {
                "description": "List domains by project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "List domains by project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Create domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Create domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vessl_dev_vessl_internal_models.DomainConfig"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/projects/{id}/summary": {
            "get": {
                "description": "GetCanvasSummary endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "GetCanvasSummary endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/services/{serviceId}/deploy": {
            "post": {
                "description": "Trigger endpoint\nTrigger Deployment",
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Backups",
                    "Deployments"
                ],
                "summary": "Trigger Deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service ID",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/services/{serviceId}/deployments": {
            "get": {
                "description": "ListServiceDeployments endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "ListServiceDeployments endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "serviceId",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/teams/{teamId}/email_settings": {
            "get": {
                "description": "Get Team Email Settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmailSettings"
                ],
                "summary": "Get Team Email Settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "description": "Save Team Email Settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmailSettings"
                ],
                "summary": "Save Team Email Settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vessl_dev_vessl_internal_models.TeamEmailSettings"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "vessl_dev_vessl_internal_models.DomainConfig": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "domainName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pathPrefix": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "redirectTo": {
                    "type": "string"
                },
                "sslCertStatus": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "vessl_dev_vessl_internal_models.TeamEmailSettings": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resendApiKey": {
                    "type": "string"
                },
                "smtpFromAddress": {
                    "type": "string"
                },
                "smtpFromName": {
                    "type": "string"
                },
                "smtpHost": {
                    "type": "string"
                },
                "smtpPassword": {
                    "type": "string"
                },
                "smtpPort": {
                    "type": "integer"
                },
                "smtpUser": {
                    "type": "string"
                },
                "teamId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "useResend": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}